requirements.txt
appdirs==1.4.3
awsebcli==3.10.1
blessed==1.14.2
botocore==1.5.52
cement==2.8.2
colorama==0.3.7
dj-database-url==0.4.2
Django==1.11
django-inline-svg==0.1.1
djangorestframework==3.6.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
gunicorn==19.7.1
httpie==0.9.9
jmespath==0.9.2
olefile==0.44
packaging==16.8
pathspec==0.5.0
Pillow==4.1.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.14.2
semantic-version==2.5.0
six==1.10.0
tabulate==0.7.5
termcolor==1.1.0
wcwidth==0.1.7
websocket-client==0.40.0
whitenoise==3.3.0
